requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
jinja2==3.1.2
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.49
langchain_mongodb==0.6.0
langchain_ollama==0.3.0
pymongo==4.11.3
motor==3.3.2
pypdf==5.4.0
python-dotenv==1.1.0
PyYAML==6.0.2
sentence-transformers==2.5.1
huggingface-hub==0.23.3